index: Number of vn that scraped not match with last vn code from vndb
When batches system implemented, the batch number will be calculated based on
difference between number of vn that registered in the vndb and
in the local database, then divide it with 10.

But the number of vn registered in the vndb will not be same as the last vn code
in the vndb.

For example, based on last vndb database statistics on may 1 2023 00:00 UTC+8,
there are 42491 vns that stored in vndb and the last code of vn is 44226,
so there is a difference of 1735. And if we generate batches based on
number of vns in vndb then 4250 batches will be formated and all of
batch will be have size of 10 codes, except last batch will be have
remains of divisior of 10.

If all the vn code that will be generated for each of batches is sequential,
then the last code that is formed is 42491 instend of 44226.

To fix this problem, when system generate vn code sequentially with the
maximum number of codes contained in each batch except for last batch,
then run scraping from the result of vn codes that has been generated.

If number of results is not same as amount of previously generated codes,
then regenerate the next codes with the amount of code generated being is
the difference between the amount of code generated previously and
the amount of code that was succesfully scraped. Do this until the
amount of scraping result will same as amount of previously
generated codes. Then store scraping result to database.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>